chore: move shared function into util folder
