+++ src/visualizations/config/adapters/dhis_highcharts/title/index.js
@@ -1,5 +1,5 @@
 import isString from 'd2-utilizr/lib/isString'
-import getFilterText from '../getFilterText'
+import getFilterText from '../../../../../util/getFilterText'
 import {
     CHART_TYPE_YEAR_OVER_YEAR_LINE,
     CHART_TYPE_YEAR_OVER_YEAR_COLUMN,
