+++ src/visualizations/config/adapters/dhis_highcharts/title/yearOverYear.js
@@ -1,4 +1,4 @@
-import getFilterText from '../getFilterText'
+import getFilterText from '../../../../../util/getFilterText'
 
 export default function(layout, metaData, dashboard) {
     const titleFragments = []
