+++ src/visualizations/util/__tests__/getFilterText.spec.js
