+++ src/visualizations/util/getFilterText.js
@@ -1,8 +1,8 @@
-import { ouIdHelper } from '../../../../modules/ouIdHelper'
-import { dimensionIs } from '../../../../modules/layout/dimensionIs'
-import { DIMENSION_ID_ORGUNIT } from '../../../../modules/fixedDimensions'
-import { dimensionGetItems } from '../../../../modules/layout/dimensionGetItems'
-import { getOuLevelAndGroupText } from '../../../../modules/getOuLevelAndGroupText'
+import { ouIdHelper } from '../../modules/ouIdHelper'
+import { dimensionIs } from '../../modules/layout/dimensionIs'
+import { DIMENSION_ID_ORGUNIT } from '../../modules/fixedDimensions'
+import { dimensionGetItems } from '../../modules/layout/dimensionGetItems'
+import { getOuLevelAndGroupText } from '../../modules/getOuLevelAndGroupText'
 
 export default function(filters, metaData) {
     if (!Array.isArray(filters) || !filters.length) {
